Extract MemberPage to its own file
This step is required so that scraper_test can be run against MemberPage
